+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * [Fix] Remove force deleted snippets from dependent snippet's `with` (#717)
 * [Fix] Comments added in SQL query to be stripped before saved as snippet (#886)
+* [Fix] Fixed bug passing :NUMBER while string slicing in query (#901)
 
 ## 0.10.2 (2023-09-22)